feat: Make automated year ranges more flexible by respecting start and end year options
This change makes it so that you can set either of start_year or
end_year in your license config and what you set there will be respected
when using automated year ranges.

Such that if you want to have the start_year be automated but use
a custom end year or end message you can do so like this:

```
    use_automated_year_ranges: true
    end_year: " and all future years"
```

Which for a file with a git created date of 2019 would result in
a Copyright year like this:

Copyright (c) 2019 and all future years

This change is backwards compatible if you're using without those
settings set. It is backwards incompatible if you were relying on the
old behaviour of those settings being ignored.